backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
python-dotenv==1.0.0
openai==1.3.7
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-timeout==2.2.0
httpx==0.25.2
slowapi==0.1.8
